Added functions to test remove and update ratings
